Primary key tests with input/output files
