+++ generator/test/generator_test_io/primary_key_always_indexed.dart
@@ -1,0 +1,10 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'primary_key_always_indexed.g.dart';
+
+@RealmModel()
+class _Questionable {
+  @PrimaryKey()
+  @Indexed()
+  late int primaryKeysAreAlwaysIndexed;
+}
