+++ generator/test/generator_test_io/primary_key_not_be_nullable.dart
@@ -1,0 +1,9 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'primary_key_not_be_nullable.g.dart';
+
+@RealmModel()
+class _Bad {
+  @PrimaryKey()
+  int? nullableKeyNotAllowed;
+}
