+++ generator/test/generator_test_io/primary_key_not_be_nullable.log
@@ -1,0 +1,13 @@
+Primary key cannot be nullable
+
+in: asset:pkg/test/generator_test_io/primary_key_not_be_nullable.dart:8:3
+  ╷
+5 │ @RealmModel()
+6 │ class _Bad {
+  │       ━━━━ in realm model for 'Bad'
+7 │   @PrimaryKey()
+8 │   int? nullableKeyNotAllowed;
+  │   ^^^^ is nullable
+  ╵
+Consider using the @Indexed() annotation instead, or make 'nullableKeyNotAllowed' an int.
+
